refactor: extract FAQ helpers
